+++ base/src/main/java/proguard/evaluation/value/MultiTypedReferenceValue.java
@@ -23,7 +23,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
-import proguard.classfile.ClassConstants;
 import proguard.classfile.Clazz;
 
 /**